refactor: extract shared REST deps into gateway_rest_common
