+++ packages/gateway_notes_rest/src/gateway_notes_rest/deps.py
@@ -1,3 +1,17 @@
 from __future__ import annotations
 
-from gateway_oss.api.deps import get_notes_service  # noqa: F401
+from typing import Annotated
+
+from fastapi import Depends
+
+from gateway_core.client.publication import PublicationClient
+from gateway_core.client.substack import SubstackClient
+from gateway_notes.service import NotesService
+from gateway_rest_common.deps import get_publication_client, get_substack_client
+
+
+def get_notes_service(
+    pub: Annotated[PublicationClient, Depends(get_publication_client)],
+    sub: Annotated[SubstackClient, Depends(get_substack_client)],
+) -> NotesService:
+    return NotesService(pub, sub)
