+++ packages/gateway_rest_common/src/gateway_rest_common/__init__.py
@@ -1,0 +1,1 @@
+from __future__ import annotations
