+++ pyproject.toml
@@ -28,6 +28,7 @@ gateway_core = { workspace = true }
 gateway_notes = { workspace = true }
 gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
+gateway_rest_common = { workspace = true }
 
 
 [tool.ruff]
@@ -39,7 +40,7 @@ select = ["E", "F", "I", "UP"]
 ignore = ["E501"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["gateway_core", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_oss", "substack_gateway"]
+known-first-party = ["gateway_core", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_oss", "gateway_rest_common", "substack_gateway"]
 
 [tool.ruff.format]
 quote-style = "double"
